separate input types and include start ActionState in InputMessage
