+++ lightyear/src/shared/input/leafwing.rs
@@ -5,9 +5,8 @@ use leafwing_input_manager::prelude::ActionState;
 
 use crate::client::config::ClientConfig;
 use crate::client::input::leafwing::LeafwingInputConfig;
-use crate::inputs::leafwing::InputMessage;
 use crate::prelude::client::ComponentSyncMode;
-use crate::prelude::{AppComponentExt, ChannelDirection, LeafwingUserAction};
+use crate::prelude::{AppComponentExt, ChannelDirection, InputMessage, LeafwingUserAction};
 use crate::protocol::message::AppMessageInternalExt;
 use crate::protocol::message::MessageType;
 use crate::server::config::ServerConfig;
